requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
tqdm>=4.65.0

# AI/ML dependencies
langchain>=0.0.348
openai>=1.3.8
tiktoken>=0.5.2

# PDF processing
PyPDF2>=3.0.1

# Vector database
faiss-cpu>=1.7.4

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# YouTube transcripts
youtube-transcript-api>=0.6.0

# Optional: Development dependencies
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0